requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Django==2.0.10
django-crispy-forms==1.9.1
django-embed-video==1.3.2
idna==2.9
Pillow==7.1.2
pytz==2020.1
requests==2.23.0
urllib3==1.25.9
